@@ -26,9 +26,9 @@
     "postcss-loader": "2.0.8",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.5.0",
+    "react": "^16.8.6",
     "react-dev-utils": "^5.0.1",
-    "react-dom": "^16.5.0",
+    "react-dom": "^16.8.6",
     "resolve": "1.6.0",
     "source-map-loader": "^0.2.1",
     "style-loader": "0.19.0",
@@ -45,12 +45,10 @@
     "webpack-dev-server": "2.9.4",
     "webpack-manifest-plugin": "1.3.2",
     "whatwg-fetch": "2.0.3",
-    "hltc-webapi-client-shared-components": "^1.2.3",
-    "@material-ui/core": "^3.0.3",
-    "@material-ui/icons": "^3.0.1",
-    "typestyle": "^2.0.1",
-    "flux": "^3.1.3",
-    "html-webpack-inline-source-plugin": "^0.0.10"
+    "hltc-webapi-client-shared-components": "^2.0.4",
+    "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
+    "typestyle": "^2.0.1"
   },
   "scripts": {
     "start": "node scripts/start.js",
@@ -61,10 +59,9 @@
   "devDependencies": {
     "@types/jest": "^23.3.2",
     "@types/node": "^10.9.4",
-    "@types/react": "^16.4.14",
-    "@types/react-dom": "^16.0.7",
-    "@types/flux": "^3.1.8",
-    "typescript": "^3.0.3"
+    "@types/react": "^16.8.17",
+    "@types/react-dom": "^16.8.4",
+    "typescript": "^3.4.5"
   },
   "jest": {
     "collectCoverageFrom": [
@@ -108,6 +105,11 @@
       "ts-jest": {
         "tsConfigFile": "D:\\cl\\dev\\webapi_client_v4.0\\account\\tsconfig.test.json"
       }
+    }
+  },
+  "proxy": {
+    "/": {
+      "target": "http://hltc.cs.ust.hk/"
     }
   },
   "babel": {
