@@ -1,15 +1,14 @@
-import * as React from 'react';
+import React from 'react';
 import EmailInput, {validateEmail} from "hltc-webapi-client-shared-components/dist/components/EmailInput";
-import {Body1, Headline, Title} from "hltc-webapi-client-shared-components/dist/components/typographies";
 import red from "@material-ui/core/colors/red";
 import DebounceButton from "hltc-webapi-client-shared-components/dist/components/DebounceButton";
 import {style} from "typestyle";
 import TextField from "@material-ui/core/TextField/TextField";
-import Login from "../data/actions/Login";
 import Avatar from "@material-ui/core/Avatar/Avatar";
 import LockIcon from "@material-ui/icons/LockOutlined";
 import Typography from "@material-ui/core/Typography/Typography";
 import indigo from "@material-ui/core/colors/indigo";
+import {UserContext} from "../data/UserContext";
 
 const registerStyles = {
     wrap: style({marginTop: 24,display:"flex",justifyContent:"center"}),
@@ -28,82 +27,66 @@ const registerStyles = {
     })
 };
 
-interface Props{
-    style?: React.CSSProperties
-}
-interface State {
-    email:string,
-    password:string,
-    errorMsg: string
-}
-namespace Styles {
-    export const title = style({margin:"8px 0 16px 0"});
-}
-class LoginWidget extends React.Component<Props,State> {
-    constructor(props:Props) {
-        super(props);
-        this.state = {
-            email:"",
-            password:"",
-            errorMsg:"",
-        }
-    }
-    whyCannotSubmit() {
-        if(!validateEmail(this.state.email)) return "enter a valid email";
-        if(this.state.password.length <= 0) return "enter your password";
+function LoginWidget() {
+    const [email, setEmail] = React.useState<string>("");
+    const [password, setPassword] = React.useState<string>("");
+    const [errorMsg, setErrorMsg] = React.useState<string>("");
+    const {login} = React.useContext(UserContext);
+    function whyCannotSubmit() {
+        if(!validateEmail(email)) return "enter a valid email";
+        if(password.length <= 0) return "enter your password";
         return null;
     }
-    async handleLogin() {
-        if(this.whyCannotSubmit()) return;
-        this.setState({errorMsg:""});
+    async function handleLogin() {
+        if(whyCannotSubmit()) return;
+        setErrorMsg("");
         try {
-            await Login.perform(this.state.email, this.state.password);
+            await login(email, password);
         }
         catch(e) {
             switch(e.message) {
                 case "INCORRECT_CREDENTIAL":
-                    this.setState({errorMsg:"incorrect username or password."});
+                    setErrorMsg("incorrect username or password.");
                     break;
                 case "NOT_ACTIVATED":
-                    this.setState({errorMsg:"your account is not verified yet. check your verification email."});
+                    setErrorMsg("your account is not verified yet. check your verification email.");
                     break;
                 default:
-                    this.setState({errorMsg:e.message})
+                    setErrorMsg(e.message);
             }
         }
     }
-    render() {
-        return (
-        <div className={registerStyles.container} {...this.props}>
+
+    return (
+        <div className={registerStyles.container}>
             <div className={registerStyles.center}>
                 <Avatar className={registerStyles.avatar}>
                     <LockIcon />
                 </Avatar>
                 <Typography variant="headline">sign in</Typography>
             </div>
-            <EmailInput value={this.state.email} label="email"
-                        onChange={e=>this.setState({email:e.target.value})}
+            <EmailInput value={email} label="email"
+                        onChange={e=>setEmail(e.target.value)}
             />
             <div style={{height:16}}/>
-            <TextField value={this.state.password} type="password" label="password"
-                       onChange={e=>this.setState({password:e.target.value})}
-                       onKeyPress={e=>e.which===13 && this.handleLogin()}
+            <TextField value={password} type="password" label="password"
+                       onChange={e=>setPassword(e.target.value)}
+                       onKeyPress={e=>e.which === 13 && handleLogin()}
             />
-            <Body1 style={{color:red[500]}}>{this.state.errorMsg}&nbsp;</Body1>
+            <Typography style={{color:red[500]}}>{errorMsg}&nbsp;</Typography>
             <div style={{height:16}}/>
             <div className={registerStyles.buttons}>
-                <DebounceButton disabled={!!this.whyCannotSubmit()}
-                                raised color="primary"
-                                onClick={()=>this.handleLogin()}>
+                <DebounceButton disabled={!!whyCannotSubmit()}
+                                variant={"contained"} color="primary"
+                                onClick={handleLogin}>
                     sign in
                 </DebounceButton>
             </div>
             <div style={{height:16}}/>
-            <Body1><a target="_blank" href="/webapi/gui/account/register.html">forget password</a></Body1>
-            <Body1>don't have an account? <a target="_blank" href="/webapi/gui/account/register.html">sign up</a> now.</Body1>
+            <Typography><a target="_blank" href="/webapi/gui/account/register.html">forget password</a></Typography>
+            <Typography>don't have an account? <a target="_blank" href="/webapi/gui/account/register.html">sign up</a> now.</Typography>
         </div>
-        );
-    }
+    );
 }
 
 export default LoginWidget;
