@@ -1,6 +1,9 @@
-import * as React from "react";
-import * as ReactDom from "react-dom";
-import RegisterView from "./view/RegisterView";
+import React from "react";
+import ReactDom from "react-dom";
+import {ActiveUserContextProviderComponent} from "./data/ActiveUserContext";
+import ActiveUserView from "./view/ActiveUserView";
 
-ReactDom.render(<RegisterView />, document.querySelector('#root'));
+ReactDom.render(<ActiveUserContextProviderComponent>
+    <ActiveUserView/>
+</ActiveUserContextProviderComponent>, document.querySelector('#root'));
 
