@@ -1,3 +1,7 @@
-export * from "./view/withActiveUser";
-export {default as User} from "./data/User";
+export * from "./data/User";
+export * from "./data/ActiveUserContext";
+export * from "./data/UserContext";
+export * from "./data/actions/register";
 export {default as ActiveUserView} from "./view/ActiveUserView";
+export {default as LoginWidget} from "./view/LoginWidget";
+export {default as RegisterView} from "./view/RegisterView";
