@@ -1,9 +1,8 @@
 import * as React from "react";
 import {default as EmailInput, validateEmail} from "hltc-webapi-client-shared-components/dist/components/EmailInput";
-import Register from "../data/actions/Register";
+import {register} from "../data/actions/register";
 import {style} from "typestyle";
 import indigo from "@material-ui/core/colors/indigo";
-import {Body1, Headline} from "hltc-webapi-client-shared-components";
 import TextField from "@material-ui/core/TextField/TextField";
 import FormControl from "@material-ui/core/FormControl/FormControl";
 import InputLabel from "@material-ui/core/InputLabel/InputLabel";
@@ -11,6 +10,7 @@ import Input from "@material-ui/core/Input/Input";
 import FormHelperText from "@material-ui/core/FormHelperText/FormHelperText";
 import red from "@material-ui/core/colors/red";
 import WhyDisableButton from "hltc-webapi-client-shared-components/dist/components/WhyDisableButton";
+import {Typography} from "@material-ui/core";
 const styles = {
     wrap: style({marginTop: 24,display:"flex",justifyContent:"center"}),
     container: style({display:"flex", flexDirection: "column", width: 400, maxWidth: '100%'}),
@@ -67,7 +67,7 @@ export default class RegisterView extends React.Component<Props, State> {
     async handleSignUp() {
         this.setState({errorMsg:""});
         try {
-            await Register.perform(this.state.email, this.state.password);
+            await register(this.state.email, this.state.password);
             this.setState({signedUp:true});
         }
         catch(e) {
@@ -83,7 +83,7 @@ export default class RegisterView extends React.Component<Props, State> {
     renderBeforeSignUp() {
         return (
             <div className={styles.wrap}><div className={styles.container}>
-                <Headline gutterBottom>create account</Headline>
+                <Typography variant={"headline"} gutterBottom>create account</Typography>
                 <br />
                 <EmailInput
                     label="email"
@@ -109,11 +109,11 @@ export default class RegisterView extends React.Component<Props, State> {
                         {this.state.password === this.state.confirmedPassword?"":"confirm the same password"}
                     </FormHelperText>
                 </FormControl>
-                <Body1 style={{color:red[500]}}>{this.state.errorMsg}&nbsp;</Body1>
+                <Typography style={{color:red[500]}}>{this.state.errorMsg}&nbsp;</Typography>
                 <div className={styles.buttons}>
                     <WhyDisableButton
                         disableReason={this.whyCannotSubmit()}
-                        raised color="primary"
+                        variant={"contained"} color="primary"
                         onClick={()=>this.handleSignUp()}>
                         SIGN UP
                     </WhyDisableButton>
@@ -124,9 +124,9 @@ export default class RegisterView extends React.Component<Props, State> {
     renderAfterSignUp() {
         return (
             <div className={styles.wrap}><div className={styles.container}>
-                <Headline>almost done</Headline>
+                <Typography variant={"headline"}>almost done</Typography>
                 <br />
-                <Body1>we have sent a verification email to <strong>{this.state.email}</strong></Body1>
+                <Typography>we have sent a verification email to <strong>{this.state.email}</strong></Typography>
             </div></div>
         );
     }
